Extracted layer1 to its own file.
