+++ index.html
@@ -28,6 +28,7 @@
         
 
     <script type="module" src="./js/gameBoard.js"></script>
+    <script type="module" src="./js/layer1.js"></script>
     <script type="module" src="./js/modelMaker.js"></script>
     <script src="./js/actors.js"></script>
     <script src="./js/moveActor.js"></script>
